- Combined test files for better maintainability.
